+++ NEWS.md
@@ -1,6 +1,7 @@
 wallace 2.0.6
 =============
 - Removed `paleobioDB` (package off CRAN)
+- Removed `rgdal` (package expiring)
 - Changed Maintainer e-mail
 
 wallace 2.0.5